data_backup automatically.

+++ factory_test_stations/test_station/test_station_seacliff_offaxis.py
@@ -170,20 +170,33 @@ class SeacliffOffAxisStation(test_station.TestStation):
                 time.sleep(0.5)
                 continue
 
-            uut_raw_dir = [(c, os.path.getctime(os.path.join(raw_dir, c))) for c in os.listdir(raw_dir)
-                           if os.path.isdir(os.path.join(raw_dir, c)) and c not in ex_file_list]
-            # backup all the raw data which is created about 8 hours ago.
-            uut_raw_dir_old = [c for c, d in uut_raw_dir if time.time() - d > 3600 * 8]
-            if len(uut_raw_dir_old) <= 0:
-                time.sleep(1)
-                continue
-            n1 = uut_raw_dir_old[-1]
-            try:
-                self.data_backup(os.path.join(raw_dir, n1), os.path.join(os.path.join(bak_dir, n1)))
-                shutil.rmtree(os.path.join(raw_dir, n1))
-            except Exception as e:
-                ex_file_list.append(n1)
-                self._operator_interface.print_to_console(f'Fail to backup file to {bak_dir}. Exp = {str(e)}')
+                cur_time = datetime.datetime.now().hour + datetime.datetime.now().minute / 60
+                if not (any([c1 <= cur_time <= c2 for c1, c2 in self._station_config.DATA_CLEAN_SCHEDULE]) and \
+                        os.path.exists(raw_dir)):
+                    time.sleep(0.5)
+                    continue
+
+                # backup all the data automatically
+                if not (os.path.exists(bak_dir) and os.path.exists(raw_dir) and os.path.samefile(bak_dir, raw_dir)):
+                    time.sleep(0.5)
+                    continue
+
+                uut_raw_dir = [(c, os.path.getctime(os.path.join(raw_dir, c))) for c in os.listdir(raw_dir)
+                               if os.path.isdir(os.path.join(raw_dir, c)) and c not in ex_file_list]
+                # backup all the raw data which is created about 8 hours ago.
+                uut_raw_dir_old = [c for c, d in uut_raw_dir if
+                                   time.time() - d > self._station_config.DATA_CLEAN_SAVED_MINUTES]
+                if len(uut_raw_dir_old) <= 0:
+                    time.sleep(1)
+                    continue
+                n1 = uut_raw_dir_old[-1]
+                try:
+                    self.data_backup(os.path.join(raw_dir, n1), os.path.join(os.path.join(bak_dir, n1)))
+
+                    shutil.rmtree(os.path.join(raw_dir, n1))
+                except Exception as e:
+                    ex_file_list.append(n1)
+                    self._operator_interface.print_to_console(f'Fail to backup file to {bak_dir}. Exp = {str(e)}')
 
     def close(self):
         self._closed = True
@@ -406,15 +419,23 @@ class SeacliffOffAxisStation(test_station.TestStation):
                     self._is_screen_on_by_op = True
                     ready = True
 
-                ready_status = self._fixture.is_ready()
-                if ready_status is not None:
-                    if ready_status == 0x00:  # load DUT automatically and then screen on
+                err_msg_list = {
+                    (0, 1): 'Axis Z running error',
+                    (0, 2): 'This command is cancel by operator',
+                    (0, 3): 'DUT transfer is timeout',
+                    (0, 4): 'Signal from sensor (Door/Grating) is error',
+                    (0, 5): 'Signal from sensor (pressing plate) is not triggered',
+                }
+                key_ret_info = self._fixture.is_ready()
+                if isinstance(key_ret_info, tuple):
+                    ready_key, ready_code = key_ret_info
+                    if ready_key == 0x00 and ready_code == 0:  # load DUT automatically and then screen on
                         ready = True  # Start to test.
                         self._is_screen_on_by_op = True
                         if not power_on_trigger:
                             self._the_unit.screen_on()
 
-                    elif ready_status in [0x03, 0x01]:
+                    elif ready_key in [0x03, 0x01] and ready_code == 0:
                         self._operator_interface.print_to_console(f'Try to change DUT status. {power_on_trigger}\n')
                         if not power_on_trigger:
                             self._the_unit.screen_on()
@@ -428,6 +449,17 @@ class SeacliffOffAxisStation(test_station.TestStation):
                             self._fixture.button_disable()
                             self._fixture.power_on_button_status(True)
                             timeout_for_dual = time.time()
+
+                    elif key_ret_info in err_msg_list:
+                        self._operator_interface.print_to_console(f'Please note: {err_msg_list[key_ret_info]}.\n', 'red')
+                        if ready_key == 0 and ready_code == 4:
+                            while ready_code != 0:
+                                self._operator_interface.operator_input(
+                                    'Hint', err_msg_list.get(key_ret_info), msg_type='warning', msgbtn=0)
+                                ready_code = self._fixture.reset()
+                    else:
+                        self._operator_interface.print_to_console(f'Recv command not handled: {key_ret_info} \n')
+
                 time.sleep(0.01)
                 tm_current = time.time()
         except Exception as e:
